scripts: move scripts to libexec
