+++ cmake/config.h.in
@@ -36,6 +36,7 @@
 #cmakedefine SCR_BGQ
 
 #define X_BINDIR "@CMAKE_INSTALL_FULL_BINDIR@"
+#define X_LIBEXECDIR "@CMAKE_INSTALL_FULL_LIBEXECDIR@"
 #define X_DATADIR "@CMAKE_INSTALL_FULL_DATADIR@"
 #define X_INCLUDEDIR "@CMAKE_INSTALL_FULL_INCLUDEDIR@"
 #define X_LIBDIR "@CMAKE_INSTALL_FULL_LIBDIR@"
