+++ scripts/python/CMakeLists.txt
@@ -4,6 +4,6 @@ SET(PYFE_TOP
   MANIFEST.in
   setup.py
 )
-INSTALL(FILES ${PYFE_TOP} DESTINATION ${CMAKE_INSTALL_BINDIR}/scrpy)
+INSTALL(FILES ${PYFE_TOP} DESTINATION ${CMAKE_INSTALL_LIBEXECDIR}/python)
 ADD_SUBDIRECTORY(scrjob)
 ADD_SUBDIRECTORY(tests)
