+++ scripts/python/scrjob/CMakeLists.txt
@@ -10,29 +10,35 @@ FOREACH(pyscript IN ITEMS ${PYFE})
 	CONFIGURE_FILE(${pyscript} ${pyscript} COPYONLY)
 ENDFOREACH(pyscript IN ITEMS ${PYFE})
 
-INSTALL(FILES ${PYFE} DESTINATION ${CMAKE_INSTALL_BINDIR}/scrpy/scrjob)
+INSTALL(FILES ${PYFE} DESTINATION ${CMAKE_INSTALL_LIBEXECDIR}/python/scrjob)
+
+SET(PYFE_BIN
+  scr_halt.py
+  scr_list_down_nodes.py
+  scr_prerun.py
+  scr_postrun.py
+  scr_run.py
+  scr_srun.py
+  scr_jsrun.py
+  scr_lrun.py
+  scr_mpirun.py
+)
 
 SET(PYFE_EXEC
+  list_down_nodes.py
+  prerun.py
+  postrun.py
   scr_check_node.py
   scr_common.py
   scr_env.py
   scr_flux.py
   scr_glob_hosts.py
-  scr_halt.py
   scr_inspect.py
-  scr_jsrun.py
   scr_kill_jobstep.py
   scr_list_dir.py
-  scr_list_down_nodes.py
-  scr_lrun.py
-  scr_mpirun.py
   scr_param.py
-  scr_postrun.py
   scr_poststage.py
-  scr_prerun.py
-  scr_run.py
   scr_scavenge.py
-  scr_srun.py
   scr_test_runtime.py
   scr_watchdog.py
 )
@@ -41,15 +47,19 @@ FOREACH(pyscript IN ITEMS ${PYFE_EXEC})
 	CONFIGURE_FILE(${pyscript} ${pyscript} COPYONLY)
 ENDFOREACH(pyscript IN ITEMS ${PYFE_EXEC})
 
+INSTALL(FILES ${PYFE_BIN}
+PERMISSIONS OWNER_EXECUTE OWNER_WRITE OWNER_READ GROUP_EXECUTE GROUP_READ WORLD_EXECUTE WORLD_READ
+        DESTINATION ${CMAKE_INSTALL_BINDIR})
+
 INSTALL(FILES ${PYFE_EXEC}
 PERMISSIONS OWNER_EXECUTE OWNER_WRITE OWNER_READ GROUP_EXECUTE GROUP_READ WORLD_EXECUTE WORLD_READ
-        DESTINATION ${CMAKE_INSTALL_BINDIR}/scrpy/scrjob)
+        DESTINATION ${CMAKE_INSTALL_LIBEXECDIR}/python/scrjob)
 
 CONFIGURE_FILE(scr_const.py.in scr_const.py @ONLY)
 SET_SOURCE_FILES_PROPERTIES(${CMAKE_CURRENT_BINARY_DIR}/scr_const.py PROPERTIES GENERATED TRUE)
 INSTALL(FILES ${CMAKE_CURRENT_BINARY_DIR}/scr_const.py
 PERMISSIONS OWNER_EXECUTE OWNER_WRITE OWNER_READ GROUP_EXECUTE GROUP_READ WORLD_EXECUTE WORLD_READ
-        DESTINATION ${CMAKE_INSTALL_BINDIR}/scrpy/scrjob)
+        DESTINATION ${CMAKE_INSTALL_LIBEXECDIR}/python/scrjob)
 
 ADD_SUBDIRECTORY(cli)
 ADD_SUBDIRECTORY(resmgrs)
