+++ scripts/python/scrjob/cli/CMakeLists.txt
@@ -10,4 +10,4 @@ FOREACH(pyscript IN ITEMS ${SCRCLI})
 	CONFIGURE_FILE(${pyscript} ${pyscript} COPYONLY)
 ENDFOREACH(pyscript IN ITEMS ${SCRCLI})
 
-INSTALL(FILES ${SCRCLI} DESTINATION ${CMAKE_INSTALL_BINDIR}/scrpy/scrjob/cli)
+INSTALL(FILES ${SCRCLI} DESTINATION ${CMAKE_INSTALL_LIBEXECDIR}/python/scrjob/cli)
