+++ scripts/python/scrjob/launchers/CMakeLists.txt
@@ -14,4 +14,4 @@ FOREACH(pyscript IN ITEMS ${JOBLAUNCHERS})
 	CONFIGURE_FILE(${pyscript} ${pyscript} COPYONLY)
 ENDFOREACH(pyscript IN ITEMS ${JOBLAUNCHERS})
 
-INSTALL(FILES ${JOBLAUNCHERS} DESTINATION ${CMAKE_INSTALL_BINDIR}/scrpy/scrjob/launchers)
+INSTALL(FILES ${JOBLAUNCHERS} DESTINATION ${CMAKE_INSTALL_LIBEXECDIR}/python/scrjob/launchers)
