+++ scripts/python/scrjob/resmgrs/CMakeLists.txt
@@ -14,4 +14,4 @@ FOREACH(pyscript IN ITEMS ${RESOURCEMANAGERS})
 	CONFIGURE_FILE(${pyscript} ${pyscript} COPYONLY)
 ENDFOREACH(pyscript IN ITEMS ${RESOURCEMANAGERS})
 
-INSTALL(FILES ${RESOURCEMANAGERS} DESTINATION ${CMAKE_INSTALL_BINDIR}/scrpy/scrjob/resmgrs)
+INSTALL(FILES ${RESOURCEMANAGERS} DESTINATION ${CMAKE_INSTALL_LIBEXECDIR}/python/scrjob/resmgrs)
