+++ scripts/python/tests/CMakeLists.txt
@@ -4,7 +4,7 @@ SET(PYFE_TESTS_TXT
   printer.c
 )
 INSTALL(FILES ${PYFE_TESTS_TXT}
-        DESTINATION ${CMAKE_INSTALL_BINDIR}/scrpy/tests)
+        DESTINATION ${CMAKE_INSTALL_LIBEXECDIR}/python/tests)
 
 SET(PYFE_TESTS_EXE
   fluxenv.sh
@@ -20,4 +20,4 @@ SET(PYFE_TESTS_EXE
 )
 INSTALL(FILES ${PYFE_TESTS_EXE}
 PERMISSIONS OWNER_EXECUTE OWNER_WRITE OWNER_READ GROUP_EXECUTE GROUP_READ WORLD_EXECUTE WORLD_READ
-        DESTINATION ${CMAKE_INSTALL_BINDIR}/scrpy/tests)
+        DESTINATION ${CMAKE_INSTALL_LIBEXECDIR}/python/tests)
